requirements.txt
python-dotenv
Flask-Cors
Jinja2
requests
openai
azure-storage-blob
datetime